Bug 1241066 - fix mozRtt always 0 or 1;r=jib My shortest patch to date.
MozReview-Commit-ID: 8r3ZrGUk40D

UltraBlame original commit: 7784662797115c9befe0f09bea2013a843740645

+++ media/webrtc/trunk/webrtc/modules/rtp_rtcp/source/rtcp_receiver.cc
@@ -568,7 +568,7 @@ void RTCPReceiver::HandleReportBlock(
                      reportBlock->lastReceivedRRNTPfrac);
 
   
-  uint32_t receiveTimeMS = Clock::NtpToMs(reportBlock->lastReceivedRRNTPsecs,
+  uint64_t receiveTimeMS = Clock::NtpToMs(reportBlock->lastReceivedRRNTPsecs,
                                           reportBlock->lastReceivedRRNTPfrac);
